Fix interactive update race condition
We first need to wait for the stack to begin updating before we start
checking if it's complete/failed. If we don't do that we can hit a race
condition between the CLI and Heat where CLI thinks that the update has
finished, but in fact it hasn't even started yet.

Closes-Bug: #1589986
Change-Id: I755fe02155a37698f8c9c885d7b6bea984e21a75