Better namespace the list of TripleO actions
The action list is quite small, but growing rapidly and already
getting confusing. In the Python package we have more detailed
namespaces to group the actions - we should use the same to group
them for users too.

This adds the new entry points, but leaves the existing names
in place. These can be removed once the CLI and GUI have
transitioned over.

Change-Id: I247747c7d641a70f320faf109c15745d00d62ff3