Verify that neither the Swift container or Mistral environment exist
Without this check it is possible for create plan to get into a
partial state. If the Mistral environment exists, but the swift
container doesn't then it will create the container and error
when attempting to create the environment. There is still a
possibility of a race condition, but this verification greatly
reduces that risk.

Change-Id: I18dee67c5935248a32b3b90241e3bbdd02034677
Closes-Bug: 1613196