Ironic Client: Bump the max_retries and retry_interval
This patch is bumping the max_retries and retry_interval when
instantiating the Ironic client to 60 seconds. This is to workaround the
NodeLocked problem that might happen to some baremetal commands.

Looking at some TripleO-CI logs it's not uncommon to see this exception
being raised when it's registering nodes in Ironic, that's because
commands like --register-nodes does also change the node's state to move
it from "manageable" to "available" which may conflict with the periodic
task in the Ironic conductor syncing the power state.

A FIXME was left in the code noting that we should use the
ironicclient.get_client() functions instead of instantiating the
Client() class directly, that's because the Client() class does not
validate any parameters.

This patch also adds a new unittest module for the actions/base.py which
is currently not unittested at all.

Closes-Bug: #1612622
Change-Id: I09501ee989d881b727434275fc0384fb5b67aa32