Set Deployment Parameters
Previously in python-tripleoclient, they tried to use UpdateIdentifier
or two different things: tell whether to perform package update, and
also to tell whether the top-level stack is being created or updated
(which was incorrect and resulted in bug 1567384, and an attempt to
work around that bug resulted in bug 1567385).

Heat's "action" conditionals cannot be used in some cases, because they
refer to the direct parent stack, which can yield undesirable results
when introducing new nested stacks or temporarily no-opping something
and then adding it back (in both these cases, "action" would be
considered "CREATE", even though the top-level stack is in "UPDATE").

UpdateIdentifier is cleared on deploy/scale to make sure that package
update doesn't get triggered, and we pass the StackAction parameter to
allow us to correctly distinguish between create vs. update of the
top-level stack, and condition the pacemaker restarts on this variable
instead.

This patch sets the parameters mentioned above and stores them in a
Mistral Environment.  These parameters are retrieved during the
ProcessTemplatesAction and subsequently merged into the Heat
environment.

Co-Authored-By: Dougal Matthews <dougal@redhat.com>
Co-Authored-By: Jeff Peeler <jpeeler@redhat.com>
Closes-Bug: #1621099
Change-Id: I88b6c1c60b2cc5de8bcf67348458c8de025779da