Return the result of create_plan in create_deployment_plan workflow
This was a regression in Iff88d1a74cf608885eed92077f85accabf1e672e
which changed the result of the workflow from task(create_plan).result to
task(process_templates).result.

The return value of a workflow should not return a totally different result
because we have added a new task at the end, this is part of our interface and
shouldn't be changed without consideration.

Closes-Bug: #1625478
Change-Id: Ib351af78df7dff35c1a44b4d0329c684358908fa