Fix the validation ssh keys workflow
The current workflow did not end up copying the validation keys to the
overcloud nodes, because the overcloud deployment action is
asynchronous. So copy_ssh_keys runs right after the Heat call when the
list of Nova servers is empty.

So instead, we optionally pass the validation key into Heat directly
at deployment time, not after.

Change-Id: I5ceee2a1db077167253f7b3b7d6160f1efd0e447
Closes-Bug: #1635226
Depends-On: I861b9e2252a9c8122dcf7df261386f1ea5200c4f