Workflow for applying RAID configuration to Ironic nodes
Introduces a new workflow for creating RAID configuration:
tripleo.baremetal.v1.create_raid_configuration using the manual
cleaning workflow introduced in the previous patch. Nodes are
expected to be in the "manageable" state.

Note that the possibility to actually build RAID depends on
a particular driver.

Partially implements: blueprint raid-workflow
Change-Id: Ied03c94bfd1a6dd5c9af8e87395da4619dd3bdca