Fix parameter update for custom roles
When updating parameters for roles, a fixed set of parameters based
on the standard roles (Compute, Control, Ceph-Storage, ...) was used to
name the count and flavor params. This didn't account for custom roles.

This patch changes this to derive the parameter names based on
convention, so updating count/flavor params for custom roles doesn't
break.

Change-Id: Iee172782ed8faa1eb077f43132083e5c6af67250
Closes-Bug: #1643882