Add log capturing support workflows
This change creates a set of workflows for use with support actions. The
first set of support actions supported is the ability to run sosreport
on various nodes and upload them to a swift container on the undercloud.
After they file bundles have been uploaded to the swift container, the
tripleoclient can be used to pull the files down to a local machine.

Change-Id: I47c486d14c46a653c61cfd92d9f484efe0407217
Partial-Blueprint: capture-environment-status-and-logs