Adds output for failure cases to delete plan action
When a plan was deleted, no feedback to the user was provided in
the cases where a container does not exist or exists but does not
contain metadata identifying it as a deployment plan container.

This patch adds feedback to the user in the above cases by adding
error text that will be returned in the resulting action failure.

Change-Id: I47267d283c35ef517b4ff2530fd662f81ed0dc88