Add FreeIPA enrollment environment generator
This is based on previous work [1] and it's what I've been using to
test the TLS-everywhere work.

This is a convenience tool that will generate an environment file
that sets the override in the resource registry as well as the
relevant parameters that will be needed in a TLS-everywhere deployment.

[1] https://github.com/JAORMX/freeipa-tripleo-incubator

bp tls-via-certmonger

Change-Id: If3e6cb1dde235b24a0c188eb758f9a2e6a626f69
Depends-On: Iac94b3b047dca1bcabd464ea8eed6f1220c844f1