Don't require mistralclient for password gen
Update the generate_overcloud_passwords function so
that it no longer explicitly requires mistralclient.
This is useful to the python-tripleoclient undercloud
installer which would like to re-use this function
and does not currently run mistral.

This patch also renames the function so that it is more
clear these passwords would get used by both the
undercloud and overcloud installers.

Change-Id: I87fe5b3c2237d2f5a9b3b4b97216b87c548cf78d