Sync template processing
tripleo-heat-templates and tripleo-common now both have their own code
for template processing as of Ie664c9c5f455b7320a58a26f35bc403355408d9b.

As the above commit mentions, this duplication should be refactored in
the future so these 2 processes can share the same code. In the short
term however, we need to keep these 2 in sync. This patch adds handling
of the disable_constraints logic to the tripleo-common code so that the
2 are in sync.

Partially-implements: blueprint split-stack-software-configuration
Change-Id: Id318d0bc3a6358749b7b83b16640012611f15c4f