Create or update deployment plan from git repository
This patch refactors the plan_management workbook to reduce
redundancy in creating deployment plans and adds the ability
to use templates stored in an external git repo when creating
or updating a deployment plan.  The workbook has tasks added
to the workflows that cleanup temporary files created during
the git clone operations.

A tag or branch can be specified in the url with an '@'.
Example: https://github.com/openstack/project.git@stable/newton

Implements: blueprint git-deployment-plan
Change-Id: If069ebb33ee4cb7f81308c68849ba9f969752d25