Use kwargs to pass in 'error' to a Mistral Result
This standardises how we create mistral.workflow.utils.Result instances.
Previously we used a combination of passing empty strings or None values
to pass errors only. Using kwargs makes this clearer and cleaner. This
is also means we use the Result class in the same way as the builtin
Mistral actions.

Change-Id: I5f93006a104048464321844a6df4a8bba131d0bb