Fix Swift ring deployment and Newton to Ocata upgrades
If an upgrade started without the latest newton/stable version of
tripleo-common, this currently will fail because the temporary URLs are
not created.

While debugging this I noticed that the workflow is not correct;
actually the Swift ring temporary URLs need to be created always before
starting the deployment. Without this patch, temporary URLs were not
created if validations are not run - which was the case when upgrading
the overcloud.

This needs to be fixed for Ocata even if the latest stable/newton
release has been deployed before the upgrade started, otherwise it might
just fail later on (if there are any deployments without doing the
validations).

Closes-Bug: 1669068
Change-Id: I765498c668932e3c95caeb53e4b73d00990898f4