Workflow to select nodes matching a profile/role
As a first step to enabling parameter calculation based on introspection
data, we'll need to find the available nodes matching a particular role
(which is associated with a flavor, which matches a specific profile in
ironic).

The next step is to retrieve the introspection data for each node output
from this workflow, then define another workflow that does the
transformation between introspection data and parameters.

Partially-Implements: blueprint tripleo-derive-parameters
Change-Id: I4b13f2923caa055be979eda57f9d5009960d61ef