Handle re-created resources in StackUpdateManager
Since Iaacc585b9806c08fd5a57c80c10eeb31cb0a8f71 got merged in Heat, we
can get into a state where a failed resource is recreated during update
after clearing a hook. That means that when we poll for events, we need
to check for that additional action and not limit to updates.

Change-Id: I2191a3f3df426eaa91fdd5b00299193a32efe366
Closes-Bug: #1677548