Add skip_deploy_identifier
Add a new action argument, skip_deploy_identifier to DeployStackAction.
The argument will disable setting a unique value for the
DeployIdentifier parameter, which means the SoftwareDeployment resources
in the templates will only be triggered if there is an actual change to
their configuration.  This argument can be used to avoid always applying
configuration, such as during node scale out.

Change-Id: Idb901a841846fec33d189b3c95f669b0380498ce