Rename contrib to container-images for packaging
A contrib directory would not be considered packagable, but we have an
increasing need to package overcloud_containers.yaml and
tripleo_kolla_template_overrides.j2 for image building and pulling in
CI and other environments.

This change renames contrib to container-images and adds symlinks in
contrib so that scripts[1] and documentation[2] that depend on this
location can be moved over to using packaged container-images rather
than git contrib.

[1] http://git.openstack.org/cgit/openstack/tripleo-quickstart-extras/tree/roles/overcloud-prep-containers/templates/overcloud-prep-containers.sh.j2#n33
    https://github.com/dprince/undercloud_containers/blob/master/build_kolla.sh#L47

[2] http://git.openstack.org/cgit/openstack/tripleo-docs/tree/doc/source/containers_deployment/tips_tricks.rst#n167
    http://git.openstack.org/cgit/openstack/tripleo-docs/tree/doc/source/containers_deployment/architecture.rst#n52
    http://git.openstack.org/cgit/openstack/tripleo-docs/tree/doc/source/containers_deployment/overcloud.rst#n58

Change-Id: Ifb736fb52e255260a4f1931b0b9b27206f2f1d82