Correct message format for plan_management workbook
In the plan management workbook we have a bunch of expressions
along the lines of "<% $.message or '' %>" to set a default
to the message if it's not set explicitly. However this is actually
wrong and will make the action fail with "YaqlEvaluationException:
Can not evaluate YAQL expression [expression=$.message or \'\',
error=u\'message\', data={}]\n'}". The correct way to do it is to
use <% $.get('message', '') %>, like the other workflows are.

Closes-Bug: 1686981
Change-Id: Ie6998c44eba46596b6d836b1201b782de761336a