Add a bootstrap_host_exec CMD
Adds a simple command line utility that will exist in all TripleO
containers (we may need this in a variety of places) so that we
can quickly figure out whether a command should execute on the
bootstrap node (or not).

For example:
  ./bootstrap_host_exec heat_engine heat-manage db_sync

would only execute the DB sync for heat on the bootstrap node.
Internally the script uses hieradata to determine when to run the
supplied command.

Also, adds openstack-tripleo-common to the base container image so we
can use it in a variety of container images.

Related-Bug: #1688380
Change-Id: If25f217bbb592edab4e1dde53ca99ed93c0e146c