Add the major upgrade converge in the update workbook
The update workbook aim to be the main entry point
for major and minor update.
The converge step is still needed for the major upgrade
so we need to add it in the workbook.

In a follow up patch, im planning to rename the
'package_update' to 'update' or 'upgrade'.

Change-Id: I70193dccd6e70f53a2d09d6de5f66a5b8a40f144