Pass all role data into role templates when rendering
The current model only passes the role name, but it will be necessary
to pass the whole role map in, e.g to enable defining a list of
custom networks, or other data except the name that we may need when
rendering per-role templates.  We fall back to the old approach for
backwards compatibility.

Change-Id: Id1249b78b3dd87e91d572ffa31b7a541f3cde2c7
Partially-Implements: blueprint composable-networks