Change cwd and set HOME for ansible actions
By changing the HOME dir and the working directory, we'll be able to
isolate ansible command executions from the rest of the environment.
While this is not going to completely isolate the execution, it'll allow
for having local `ansible.cfg` files, local temp dirs, and other
artifacts required to execute and customize ansible.

Change-Id: Ic67d91d6bad64f62c6b69c9627927ad8261d0555