Remove EPEL in TripleO's Kolla template_overrides
Kolla has been trying to remove EPEL as a dependency from their
base container image but it turns out it's not so simple [1] and could
potentially take until Queens.

RDO does not require EPEL since Mitaka and can only cause problems
for us in the long run so let's remove it.

[1]: https://review.openstack.org/#/c/474999/

Change-Id: Ib235810a2d2776708a352d117a4a85075424e33d