Publish logs before exporting them
TripleO UI logs are moved from a Zaqar queue to a Swift object once an hour via
a cron job.  This means that when a log download is requested from the UI, there
could be a gap of up to one hour in log coverage.  This patch calls the
`publish_ui_logs_to_swift` workflow before exporting the logs to a file; this
ensures that all available log data are present in the downloaded archive.

Change-Id: I9a34186040f41f77e42195197ea95b7dcbdcd995
Implements: blueprint websocket-logging