overcloud_containers.yaml.j2 map images to services
This change adds a services key to each entry in
overcloud_containers.yaml.j2 which contains a list of
tripleo-heat-templates docker resource types which consume this image.

The intent is for the prepare command to use this plus the heat
environments to produce an image list that only contains images for
services actually being deployed. This targeted list will make image
upload operations quicker, which is one issue affecting containers in
CI.

Change-Id: Ie8801b11921c46923b0f7c9aaba6bf524c464e82
Partial-Bug: #1710992