Discover a versioned container image tag
This change implements a discover_image_tag method to discover the
versioned tag for a container image.

The current approach is to expect a label to be set on the image which
contains the expected tag value. For example the RDO image building
pipeline sets a label rdo_version whose value also matches a tag.

This change will be used by the python-tripleoclient change which
actually implements the command
"openstack overcloud container image tag discover".

Change-Id: I27ea031287604d70032fb5392aecbce313d4b096
Partial-Bug: #1708967