Fix healthchecks permissions
Set the permissions on the files rather than recursively on the
directory otherwise symlinks are ignored.

According to the chmod man page:

"For each symbolic link listed on the command line, chmod changes the
permissions of the pointed-to  file.  In contrast, chmod ignores
symbolic links encountered during recursive directory traversals."

Closes-Bug: #1718895
Change-Id: I3c6545e0be2f70a0f7e90fbf9fda2646c36217dd