Fail ceph-ansible workflow if there is stale puppet-ceph hieradata
We should stop the deployment if there is stale puppet-ceph
hieradata because it isn't converted automatically into
ceph-ansible variables.

Initially we only stop the deployment if ceph::profile::params::osds
is found to make sure operators have converted that into a ceph-ansible
parameter and removed it from the Heat environment file.

Change-Id: I6031625b128b989c78f20350d2971cc61bf874b1
Closes-Bug: #1723085