Finish deprecation of pull_source
In change I0b0b8728b7752bdd4680a5a2bbf664f94f7d56d9 the --pull-source
argument was deprecated in favour of a --namespace which includes the
source registry.

This change makes the default namespace docker.io/tripleoupstream to
be consistent with the default that the CLI uses. Tests now do not use
pull_source so that proper testing of fully qualified namespace
arguments is done.

Change-Id: Icd29f0e75c3edadfb008056af774e283200002b0
Related-Bug: #1708303