Fix typos in container_images_prepare failure task names
The message tasks in case of failures should reference the task to
prepare the container images itself. In the case of create_plan, it
would return a message for the wrong task, and in the case of
update_plan I think it would fail altogether. This now references the
correct task name.

Change-Id: Ie05bd113b8c20ee95aef82b07c63a9591baab98a