Add ansible and pyton-shade to the base image
Ansible and shade will be used to run some of the provisioning
operations as part of the kubernetes and ansible role based deployment
of the services. Having it in the base images makes it possible to use
any of the service images to run ansible commands, just like we do with
puppet.

Change-Id: I76efb92e27955b6e9054d4e7dcfda5ba1d8b2d2f