Implement post-upload cleanup of docker images
The docker uploader will leave a copy of all uploaded images in the
local docker storage. This change will track those images and delete
them from the local docker after all uploads are complete.

If a local image is in use (for example, an already deployed
containerised undercloud) then the delete will fail. In this case,
only a warning is logged.

Change-Id: Ic0424638b9ddbf77e10cfe936d0b96ff2da1a59e
Closes-Bug: #1708965
Closes-Bug: #1694709