Always load rsyslog-base container
It used to be the case that one would need to explicitly tell the image
builder to run load the rsyslog-base container (via the stdout
parameter). And this commit removes that, and always loads this
container.

This is with the purpose of removing swift's dependency to the host's
rsyslog instance. By using a sidecar container for swift, we no longer
will need to rely on the host's rsyslog instance to actually persist
swift's logs.

bp logging-stdout-rsyslog

Change-Id: Idca722f1183eec871152b8f52554594bbf1f67e8