Include barbican worker and listener images if API is enabled
Due to limitations in the puppet manifest and the need to go forward
with the barbican work, we will trigger the usage of the worker and
listener images when the barbican API is used.

In other words, for the near future, we will require the notifier and
listener to be in the same "pod" as the barbican API.

I'm leaving the listener and worker services entries there because we do
hope to separate these at some point.

Change-Id: I154450c96e25a045a16cba23eb7fd28e309a9a60