Add puppet-tripleo & python2-kubernetes
Thinking ahead to a kubernetes based deployment, it would be great to
have these 2 packages in the base image for a couple of reasons:

1) We can create k8s jobs to run puppet apply and, while it is good and
possible to mount the puppet modules on development environments, it is
not as convenient for production.

2) There are a couple of ansible modules and custom scripts that we will
run from the images: Uploading generated config files to config maps,
managing resources from PODs, etc. and we'll need the kubernetes module
for that.

Eventually, we will also need python-openshift, which is not in the
repos we have enabled right now. When the time comes, I'll propose a
patch to add it.

Change-Id: I61e35d8118c1de4c2976f496e8a6c9c529f3d91f