Parameterize Ansible environment vars in Mistral Workflow
Add ansible_env_variables which a user may use to override
any Ansible environment variable. This new parameter is
then merged with the default env variables as set in
ansible_extra_env_variables.

Also adds DEFAULT_FORKS to the default env variables
with a new default of 50. Ansible automatically limits
the number of forks to the number of possible hosts at
runtime so it is not necessary to compute the fork count.

Change-Id: I5d69af146ca6ca8b3d5f78445cd1b47828daa955
Partial-Bug: 1738276