Lowering the number of ansible forks to 25
With forks set to 50, we can see upwards of 20GB RSS memory used on the
undercloud (depdning on the size of the overcloud). To reduce this, and
the chance for Ansible to run out of memory, I am recommending to update
this to 25.

To see the impact of forks set to 25:
https://snapshot.raintank.io/dashboard/snapshot/gseMLc1opbSueaY91tcjjbH9Lub2tnBm?orgId=2

Change-Id: I7c57b641aa7fea02f865778321c98d1dd3d7e085
Closes-Bug: 1740880