Add missing support for the root_device property
Our documentation requires it to be set for multi-disk nodes, but it's
not accepted in instackenv.json. This patch fixes it.

As a side fix, property values are no longer converted to strings. This
has never been required, and may potentially yield incorrect results
(like with root_device, which is a dictionary).

Change-Id: I4e9fbc940c424071206e96b3d90086282f381a9f
Closes-Bug: #1742905